Error fix
Reported on https://github.com/mchlbowyer/master-login-system/issues/2

+++ lib/captcha/captcha.php
@@ -51,7 +51,7 @@ function captcha($config = array()) {
     if( $captcha_config['max_font_size'] < $captcha_config['min_font_size'] ) $captcha_config['max_font_size'] = $captcha_config['min_font_size'];
 
     // Use milliseconds instead of seconds
-    srand(microtime() * 100);
+    srand((float) microtime() * 100);
 
     // Generate CAPTCHA code if not set by user
     if( empty($captcha_config['code']) ) {
